backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
pydantic
sqlalchemy
psycopg2-binary
